% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{dtriang}
\alias{dtriang}
\title{Triangular density function}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector of values}

\item{min}{Lower bound}

\item{max}{Upper bound}

\item{mode}{Mode of the distribution}
}
\value{
Numeric vector of densities
}
\description{
Computes the probability density function of the triangular distribution.
}
