\name{TRIANG}
\alias{TRIANG}
\alias{TRIANG_package}
\docType{package}
\title{Discrete triangular distributions}
\description{The function plots discrete triangular distributions}
\details{
\tabular{ll}{
Package: \tab PackageTRIANG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-19\cr
License: \tab GPL-2\cr
}

}
\author{
Tristan Senga Kiess\'e, Silvio S. Zocchi, C\'elestin C. Kokonendji

Maintainer: <tristan.sengakiesse@univ-pau.fr>
}
\references{
Kokonendji, C.C., Senga Kiess\'e, T. and Zocchi, S.S. (2007). Discrete
triangular distributions and non-parametric estimation for probability mass
function. Journal of Nonparametric Statistics 19, 241--254.
}
\keyword{distribution}
