% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{ptriang}
\alias{ptriang}
\title{Triangular cumulative distribution function}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector}

\item{min}{Lower bound}

\item{max}{Upper bound}

\item{mode}{Mode}
}
\value{
Numeric vector
}
\description{
Triangular cumulative distribution function
}
