% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{qtriang}
\alias{qtriang}
\title{Triangular quantile function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Numeric vector of probabilities}

\item{min}{Lower bound}

\item{max}{Upper bound}

\item{mode}{Mode}
}
\value{
Numeric vector
}
\description{
Triangular quantile function
}
