% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang.R
\name{rtriang}
\alias{rtriang}
\title{Random generation from triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations}

\item{min}{Lower bound}

\item{max}{Upper bound}

\item{mode}{Mode}
}
\value{
Numeric vector
}
\description{
Random generation from triangular distribution
}
